move ImageFields
makes deleting ES1 easier
