+++ media-api/app/lib/elasticsearch/impls/elasticsearch1/ElasticSearch.scala
@@ -1,8 +1,9 @@
 package lib.elasticsearch.impls.elasticsearch1
 
+import com.gu.mediaservice.lib.ImageFields
 import com.gu.mediaservice.lib.argo.ArgoHelpers
 import com.gu.mediaservice.lib.auth.Authentication.Principal
-import com.gu.mediaservice.lib.elasticsearch.{ElasticSearchClient, ElasticSearchConfig, ImageFields}
+import com.gu.mediaservice.lib.elasticsearch.{ElasticSearchClient, ElasticSearchConfig}
 import com.gu.mediaservice.model.{Agencies, Image}
 import com.gu.mediaservice.syntax._
 import lib.elasticsearch._
