+++ thrall/app/lib/ElasticSearch6.scala
@@ -1,6 +1,6 @@
 package lib
 
-import com.gu.mediaservice.lib.elasticsearch.ImageFields
+import com.gu.mediaservice.lib.ImageFields
 import com.gu.mediaservice.lib.elasticsearch6.{ElasticSearch6Config, ElasticSearch6Executions, ElasticSearchClient, Mappings}
 import com.gu.mediaservice.lib.formatting.printDateTime
 import com.gu.mediaservice.model._
